refactor: HomeType 모듈화
