+++ src/app/store/ducks/home/homeSlice.ts
@@ -1,4 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { HomeType } from "@type";
 
 const DUMMY_ARTICLES = [
     {
@@ -86,63 +87,10 @@ const DUMMY_ARTICLES = [
     },
 ];
 
-type storiesScrollPositionType = "left" | "right" | "center";
-
-type activatedModalType =
-    | "hover"
-    | "unfollowing"
-    | "report"
-    | "articleMenu"
-    | "shareWith"
-    | null;
-
-interface homeModalProps {
-    activatedModal: activatedModalType;
-    handledObj: null;
-}
-
-export interface homeStateProps {
-    storiesScrollPosition: storiesScrollPositionType;
-    articles: {
-        imgs: string[];
-        location: string;
-        hashtags: string[];
-        text: string;
-        owner: {
-            username: string;
-            avatarUrl: string;
-        };
-        likes: string[];
-        comments: {
-            username: string;
-            comment: string;
-        }[];
-        createdAt: number;
-    }[]; // 백엔드에서 댓글과 이 게시물에 내가 좋아요를 눌렀는지까지 보내주는지 등등
-    // request와 response에 대해 소통 필요
-    hoveredUser: {
-        avatarUrl: string;
-        verified: boolean;
-        isFollowing: boolean;
-        realName: string;
-        link: string;
-        followingUsernames: string[];
-        articlesNum: number;
-        followersNum: number;
-        followsNum: number;
-        recentImgs: {
-            src: string;
-            param: string;
-        }[]; // 최신 3개
-    } | null;
-    isCopiedNotification: boolean;
-    homeModal: homeModalProps;
-}
-
 // 좋아요 여부를 따로? 아니면 내가 필터링?
 // 좋아요 수를 따로? 아니면 내가 arr 길이 계산?
 
-const initialState: homeStateProps = {
+const initialState: HomeType.homeStateProps = {
     storiesScrollPosition: "left",
     articles: [],
     hoveredUser: null,
@@ -165,7 +113,7 @@ const homeSlice = createSlice({
         },
         changeStoriesScrollPosition: (
             state,
-            action: PayloadAction<storiesScrollPositionType>,
+            action: PayloadAction<HomeType.storiesScrollPositionType>,
         ) => {
             state.storiesScrollPosition = action.payload;
         },
@@ -173,7 +121,7 @@ const homeSlice = createSlice({
         getArticles: (state) => {
             state.articles = DUMMY_ARTICLES;
         },
-        startModal: (state, action: PayloadAction<homeModalProps>) => {
+        startModal: (state, action: PayloadAction<HomeType.homeModalProps>) => {
             state.homeModal = action.payload;
             // 비동기적으로 데이터를 가져옴
         },
